Move aligner methods into different files
