+++ metagraph/experiments/main.cpp
@@ -23,7 +23,8 @@
 #include "common/vectors/bit_vector_sd.hpp"
 #include "common/utils/template_utils.hpp"
 #include "common/data_generation.hpp"
-#include "graph/alignment/aligner_helper.hpp"
+#include "graph/alignment/aligner_config.hpp"
+#include "graph/alignment/aligner_cigar.hpp"
 #include "kmer/alphabets.hpp"
 #include "seq_io/kmc_parser.hpp"
 #include "cli/config/config.hpp"
