+++ metagraph/src/graph/alignment/aligner_cigar.cpp
@@ -1,4 +1,4 @@
-#include "aligner_helper.hpp"
+#include "aligner_cigar.hpp"
 
 #include "kmer/alphabets.hpp"
